trashclassifier/requirements.txt
# Core ML and CV libraries
numpy>=1.19.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
joblib>=1.0.0

# Image processing
Pillow>=8.0.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Optional: GPU acceleration
# tensorflow>=2.8.0
# torch>=1.10.0
